Fixing crash bug
If bluetooth is not enabled when the app starts then it crashes because
it attempts to create a Bluetooth Socket without having bluetooth
enabled. It requests the users permission but does not wait. I added the
part inside the main UI thread that will ask the user, but I have to
figure out the onActivityResult() method.